document_processing_service/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
torch==2.0.1
transformers==4.30.2
PyPDF2==3.0.1
python-docx==0.8.11
gensim==4.3.2
chromadb==0.3.21
httpx==0.23.3
